CARDS-1492: Slack notifications: # of upcoming appts, # of initial emails sent, # of reminder emails sent, # of surveys completed
Make HTTP requests using the (now introduced) io.uhndata.cards.httprequests module which makes requests using the org.apache.httpcomponents module

This io.uhndata.cards.httprequests module will serve, in the future, as the interface through which CARDS will use to make all external HTTP requests.
