+++ proms-resources/backend/pom.xml
@@ -58,6 +58,16 @@
       <version>2.0.1</version>
     </dependency>
     <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpclient-osgi</artifactId>
+      <version>4.5.6</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.httpcomponents</groupId>
+      <artifactId>httpcore</artifactId>
+      <version>4.4.10</version>
+    </dependency>
+    <dependency>
       <groupId>org.apache.sling</groupId>
       <artifactId>org.apache.sling.api</artifactId>
     </dependency>
@@ -129,6 +139,11 @@
     <dependency>
       <groupId>io.uhndata.cards</groupId>
       <artifactId>cards-performance-notifications</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.uhndata.cards</groupId>
+      <artifactId>cards-http-requests</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
